requirements.txt
dspy-ai>=2.4.0
pydantic>=2.0.0
anthropic>=0.25.0
asyncio>=3.4.3
lxml>=4.9.0